feat: Adapt predict pipeline
