--- Data - Extra/Labels-Extra.csv
@@ -1,11 +1,0 @@
-Filename,Extra
-img_25.png,Yes
-img_31.png,Yes
-img_43.png,Yes
-img_48.png,Yes
-img_90.png,Yes
-img_6001.png,No
-img_5001.png,No
-img_4001.png,No
-img_3000.png,No
-img_2000.png,No
