--- Data - Is GenAI Full/IsGenAI.txt
@@ -1,10 +1,0 @@
-Charlotte Pass - Ai.png
-Charlotte Pass - Real.png
-Dead Horse Gap - Ai.png
-Dead Horse Gap - Real.png
-Farm - Ai.png
-Farm - Real.png
-Perish - Ai.png
-Perish -Real.png
-Sunset - Ai.png
-Sunset - Real.png
