--- Data - Is Needs Respray Full/Is Needs Respray.txt
@@ -1,10 +1,0 @@
-paver weeds - 110.png
-paver weeds - 111.png
-paver weeds - 112.png
-paver weeds - 113.png
-paver weeds - 114.png
-paver weeds - 115.png
-paver weeds - 116.png
-paver weeds - 117.png
-paver weeds - 118.png
-paver weeds - 119.png
