+++ models/Is Epic/model_Is Epic.pth
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f7308db4948b1025a7588ead81fd64b8411f8ca1160688ef02f03c6d3e72bd7a
-size 14299826
+oid sha256:e741370cf1c521e09529c42edf8f534ecbc55b2e2b63cff7632882e34cbc9752
+size 14265522
