+++ models/Is GenAI/model_Is GenAI.pth
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:e4ba3511b12b62e440f3d4b145d9b00a055eb5dfc5ae8b99b40e2d8eda47d7bd
-size 14300144
+oid sha256:66a5bc6185c6c8691e5c797defa289ffc7377e6658a056be862a6ff6334c3589
+size 14265840
