+++ predictions/Is Epic Intro Full.csv
@@ -1,0 +1,11 @@
+Filenames,Is Epic
+better-day-186374.mp3.png,Yes
+calm-background-for-video-121519.mp3.png,No
+floating-yoga-meditation-background-music-149045.mp3.png,No
+lady-of-the-80x27s-128379.mp3.png,No
+youx27re-beautiful_short-2-59sec-197074.mp3.png,No
+apocalypse_v3_59sec-178285.mp3.png,Yes
+celestial-dawn-194596.mp3.png,Yes
+corporate-power-suspense-sci-fi-background-145902.mp3.png,Yes
+fight-142564.mp3.png,Yes
+warrior_30sec-192838.mp3.png,Yes
