+++ predictions/Is GenAI Full.csv
@@ -1,0 +1,11 @@
+Filenames,Is GenAI
+Charlotte Pass - Ai.png,No
+Charlotte Pass - Real.png,No
+Dead Horse Gap - Ai.png,No
+Dead Horse Gap - Real.png,No
+Farm - Ai.png,No
+Farm - Real.png,No
+Perish - Ai.png,No
+Perish - Real.png,No
+Sunset - Ai.png,Yes
+Sunset - Real.png,No
