+++ train.py
@@ -18,6 +18,8 @@ from sklearn.metrics import f1_score
 import numpy as np
 from torchvision.models import MobileNet_V2_Weights
 
+from common import Prototypical
+
 
 ########################################################################################################################
 # NOTE: Model Implementation
@@ -130,37 +132,6 @@ class CBAMAttentionMN(nn.Module):
         x = self.dense2(x)
         output = self.sigmoid(x)
         return output
-    
-    
-# Prototypical network implementation based on https://colab.research.google.com/drive/1TPL2e3v8zcDK00ABqH3R0XXNJtJnLBCd?usp=sharing#scrollTo=UW5Rxifk7Kru
-class Prototypical(nn.Module):
-    #Initialise model with resNet as base CNN and flattening fc layer
-    def __init__(self):
-        super(Prototypical, self).__init__()
-        
-        # Uncomment below to use CBAM attention model
-        # self.baseCNN = CBAMAttentionMN((224, 224))
-        
-        # self.baseCNN = models.mobilenet_v2(pretrained=True)
-        self.baseCNN = models.mobilenet_v2(weights=MobileNet_V2_Weights.IMAGENET1K_V1)
-    def forward(
-        self,
-        support_images: torch.Tensor,
-        support_labels: torch.Tensor,
-        query_images: torch.Tensor,
-    ) -> torch.Tensor:
-        support_features = self.baseCNN.forward(support_images)
-        query_features = self.baseCNN.forward(query_images)
-
-        # Infer the number of different classes from the labels of the support set
-        num_classes = len(torch.unique(support_labels))
-        # Prototype i is the mean of all instances of features corresponding to labels == i
-        prototype = torch.cat([support_features[torch.nonzero(support_labels == label)].mean(0) for label in range(num_classes)])
-
-        # Compute the euclidean distance from queries to prototypes
-        scores = -(torch.cdist(query_features, prototype))
-
-        return scores
 
 
 def episodic_evaluate(model, data_loader, criterion):
